+++ R/ggplot2_functions.R
@@ -1,6 +1,14 @@
-# the following code is copied and modified from ggplot2 under the R/aes.R file
-# (under MIT license). Implements the soft deprecated `aes_string()` function as
-# `apotc_aes_string()`
+# the following code is copied and modified from ggplot2 under the MIT license
+
+apotc_aes <- function(...) {
+    args <- Filter(Negate(rlang::quo_is_missing), enquos(...))
+    local({
+        aes <- function(...) NULL
+        rlang::inject(aes(!!!args))
+    })
+    aes <- ggplot2:::new_aes(args, env = parent.frame())
+    apotc_rename_aes(aes)
+}
 
 apotc_rename_aes <- function(x) {
   names(x) <- ggplot2::standardise_aes_names(names(x))